Removing the container when it is stopped (#1)
* Removing docker container when it is stopped

This is to avoid getting the error while we are running the demo multiple times on the same machine.

* Added build as a step on read me for newbies like me

Though it should have been straightforward, it took me few minutes to find out that I should have run build script first. So, I am editing this as an effort to make it easier for newbies.